+++ src/model/Price.java
@@ -1,5 +1,6 @@
 package model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class Price {
@@ -7,10 +8,10 @@ public class Price {
 	 * timeStamp 
 	 */
 	private LocalDateTime priceTimeStamp;
-	private double preVATValue;
+	private BigDecimal preVATValue;
 	
 	
-	public Price(double preVATValue) {
+	public Price(BigDecimal preVATValue) {
 		priceTimeStamp = LocalDateTime.now();
 		this.preVATValue = preVATValue;
 	}
@@ -19,12 +20,12 @@ public class Price {
 	 Added a second constructor, with the purpose of being able to create a price object both at the current time
 	 and at a chosen time.
 	 */
-	public Price(LocalDateTime priceTimeStamp, double preVATValue) {
+	public Price(LocalDateTime priceTimeStamp, BigDecimal preVATValue) {
 		this.priceTimeStamp = priceTimeStamp;
 		this.preVATValue = preVATValue;
 	}
 	
-	public double getPreVATValue() {
+	public BigDecimal getPreVATValue() {
 		return preVATValue;
 	}
